test: add test coverage to fs.promises.chmod and filehandle.chmod
