+++ test/lib/index.spec.js
@@ -231,91 +231,6 @@ describe('The API', function() {
 
 describe('Mocking the file system', function() {
 
-  describe('fs.chmod(path, mode, callback)', function() {
-    beforeEach(function() {
-      mock({
-        'file.txt': mock.file({mode: parseInt('0644', 8)})
-      });
-    });
-    afterEach(mock.restore);
-
-    it('changes permissions of a file', function(done) {
-      fs.chmod('file.txt', parseInt('0664', 8), function(err) {
-        if (err) {
-          return done(err);
-        }
-        const stats = fs.statSync('file.txt');
-        assert.equal(stats.mode & parseInt('0777', 8), parseInt('0664', 8));
-        done();
-      });
-    });
-
-    it('fails if file does not exist', function(done) {
-      fs.chmod('bogus.txt', parseInt('0664', 8), function(err) {
-        assert.instanceOf(err, Error);
-        done();
-      });
-    });
-  });
-
-  describe('fs.chmodSync(path, mode)', function() {
-    beforeEach(function() {
-      mock({
-        'file.txt': mock.file({mode: parseInt('0666', 8)})
-      });
-    });
-    afterEach(mock.restore);
-
-    it('changes permissions of a file', function() {
-      fs.chmodSync('file.txt', parseInt('0644', 8));
-      const stats = fs.statSync('file.txt');
-      assert.equal(stats.mode & parseInt('0777', 8), parseInt('0644', 8));
-    });
-
-    it('fails if file does not exist', function() {
-      assert.throws(function() {
-        fs.chmodSync('bogus.txt', parseInt('0644', 8));
-      });
-    });
-  });
-
-  describe('fs.fchmod(fd, mode, callback)', function() {
-    beforeEach(function() {
-      mock({
-        'file.txt': mock.file({mode: parseInt('0666', 8)})
-      });
-    });
-    afterEach(mock.restore);
-
-    it('changes permissions of a file', function(done) {
-      const fd = fs.openSync('file.txt', 'r');
-      fs.fchmod(fd, parseInt('0644', 8), function(err) {
-        if (err) {
-          return done(err);
-        }
-        const stats = fs.statSync('file.txt');
-        assert.equal(stats.mode & parseInt('0777', 8), parseInt('0644', 8));
-        done();
-      });
-    });
-  });
-
-  describe('fs.fchmodSync(fd, mode)', function() {
-    beforeEach(function() {
-      mock({
-        'file.txt': 'content'
-      });
-    });
-    afterEach(mock.restore);
-
-    it('changes permissions of a file', function() {
-      const fd = fs.openSync('file.txt', 'r');
-      fs.fchmodSync(fd, parseInt('0444', 8));
-      const stats = fs.statSync('file.txt');
-      assert.equal(stats.mode & parseInt('0777', 8), parseInt('0444', 8));
-    });
-  });
-
   describe('fs.unlink(path, callback)', function() {
     beforeEach(function() {
       mock({
